Break out Mask type to mask.rs
